python exercises

+++ Angela_Python/calculator.py
@@ -29,14 +29,14 @@ operations = {
 
 
 def calculator():
-    num1 = int(input("What's the first number?: "))
+    num1 = float(input("What's the first number?: "))
     for symbol in operations:
         print(symbol)
     should_continue = True
 
     while should_continue:
         operation_symbol = input("Pick an operation from the line above: ")
-        num2 = int(input("What's the second number?: "))
+        num2 = float(input("What's the second number?: "))
         calculation_function = operations[operation_symbol]
         answer = calculation_function(num1, num2)
